fix issues emun
