+++ src/typeScript/controllers/todoController.ts
@@ -1,4 +1,4 @@
-import { ACTION_FORM, KEY, PAGE } from "../constants/enum";
+import { ACTION_FORM, KEY, PAGE } from "../enums";
 import { ITodo, ITodoParam } from "../constants/interface";
 import { handleCreateId } from "../helper/createId";
 import { getLocalStorage, clearLocalStorage } from "../helper/localStorage";
