+++ src/typeScript/enums/form.ts
@@ -12,22 +12,6 @@ export enum FORM_EVEN {
   BLUR = "blur",
   FOCUS = "focus",
 }
-
-export enum KEY {
-  LOCALSTORAGE_UESR = "user",
-  LOCALSTORAGE_TODOS = "todos",
-  LOCALSTORAGE_ID_UPDATE = "idUpdate",
-}
-export enum TOAST {
-  SUCCESS = "success",
-  ERROR = "error",
-}
-
-export enum PAGE {
-  LOGIN = "loginPage",
-  TODO = "todoPage",
-}
-
 export enum ACTION_FORM {
   ADD = "ADD",
   UPDATE = "UPDATE",
