+++ src/typeScript/enums/index.ts
@@ -1,0 +1,6 @@
+import { FORM, FORM_EVEN, ACTION_FORM } from "./form";
+import { KEY } from "./localstorage";
+import { PAGE } from "./page";
+import { TOAST } from "./toast";
+
+export { PAGE, FORM, FORM_EVEN, ACTION_FORM, TOAST, KEY };
