+++ src/typeScript/enums/localstorage.ts
@@ -1,0 +1,5 @@
+export enum KEY {
+  LOCALSTORAGE_UESR = "user",
+  LOCALSTORAGE_TODOS = "todos",
+  LOCALSTORAGE_ID_UPDATE = "idUpdate",
+}
