+++ src/typeScript/enums/page.ts
@@ -1,0 +1,4 @@
+export enum PAGE {
+  LOGIN = "loginPage",
+  TODO = "todoPage",
+}
