+++ src/typeScript/enums/toast.ts
@@ -1,0 +1,4 @@
+export enum TOAST {
+  SUCCESS = "success",
+  ERROR = "error",
+}
