+++ src/typeScript/views/todoView.ts
@@ -1,4 +1,4 @@
-import { ACTION_FORM, KEY } from "../constants/enum";
+import { ACTION_FORM, KEY } from "../enums";
 import { ITodo } from "../constants/interface";
 import debounce from "../helper/debounce";
 import { handleFormTodo } from "../helper/form";
